+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix #5218: No export for `io.fabric8.tekton.triggers.internal.knative.pkg.apis.duck.v1beta1` in tekton v1beta1 triggers model
 * Fix #5224: Ensuring jetty sets the User-Agent header
 * Fix #4225: Enum fields written in generated crd yaml
+* Fix #5236: using scale v1beta1 compatible logic for DeploymentConfig
 
 #### Improvements